+++ parm/md_ext.py
@@ -1,6 +1,5 @@
-from __future__ import absolute_import
-from __future__ import unicode_literals
-from future.builtins import range, open
+from __future__ import absolute_import, unicode_literals
+from ._compat import range, open
 from .utils import log
 import re
 import os